more refactoring and actual stable release

+++ src/wbs/platform/php/AbstractPhpEntity.java
@@ -133,7 +133,7 @@ class AbstractPhpEntity
 
 	@Override
 	public
-	Integer asInteger () {
+	Long asInteger () {
 
 		throw new PhpTypeException (
 			type,
